[feat] reduce product amount when order sent via post

+++ src/main/java/com/shop/dto/ShopCardDto.java
@@ -22,5 +22,5 @@ public class ShopCardDto {
 
     private String imageUrl;
 
-    private Integer inCardAmount;
+    private Long inCardAmount;
 }
